// clients/trait-wars/types/effects.ts
export type CombatActionType =
    | 'melee'
    | 'ranged'
    | 'magic'
    | 'heal'
    | 'defend'
    | 'hit'
    | 'death'
    | 'buff';